Revert "Move the interfaces for Volume, Page and Section to the API and make the getVolume function public."
This reverts commit 68e7b4d6dcf19c47e777b12e3ac2979bb35a37cc.
