+++ src/org/daisy/dotify/formatter/impl/page/PageImpl.java
@@ -12,7 +12,7 @@ import org.daisy.dotify.api.translator.DefaultTextAttribute;
 import org.daisy.dotify.api.translator.Translatable;
 import org.daisy.dotify.api.translator.TranslationException;
 import org.daisy.dotify.api.writer.Row;
-import org.daisy.dotify.api.formatter.Page;
+import org.daisy.dotify.formatter.impl.common.Page;
 import org.daisy.dotify.formatter.impl.core.BorderManager;
 import org.daisy.dotify.formatter.impl.core.FormatterContext;
 import org.daisy.dotify.formatter.impl.core.HeightCalculator;
