+++ src/org/daisy/dotify/formatter/impl/sheet/VolumeImpl.java
@@ -1,7 +1,7 @@
 package org.daisy.dotify.formatter.impl.sheet;
 
-import org.daisy.dotify.api.formatter.Section;
-import org.daisy.dotify.api.formatter.Volume;
+import org.daisy.dotify.formatter.impl.common.Section;
+import org.daisy.dotify.formatter.impl.common.Volume;
 import org.daisy.dotify.formatter.impl.search.Overhead;
 
 import java.util.ArrayList;
